#7 / Refactored code base and splittet RemoteReplicaActor
The RemoteReplicaActor is splittet into Receiver and sender, were the
sender holds the replicas session key. For every external replica there
is a receiver with its corresponding session key.

The code base is restructured and refactored to be more testable
without mocks.

Right now none of the tests are working. This will be repaired with the
upcoming changes.